requirements.txt
# Habitz unified app — combined requirements from all four trackers
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
WTForms==3.1.1
Werkzeug==3.1.3
email-validator==2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn==23.0.0
anthropic>=0.25.0
